Add files via upload
I've made some changes to the script based on the feedback we received. First, I updated the script to save the system information file to the current user's desktop instead of the root directory. This should make it easier for users to find the file after it's been created.

Secondly, I added some code to create a custom power plan based on the Balanced power plan and set some specific settings for screen saver timeout, sleep timeout, and system sleep behavior. This should help ensure that the computer is set up with the optimal power settings for our use case.

Finally, I updated the GUID used to create the custom power plan to a new one that we generated. This was necessary because the old GUID was specific to a different machine and could cause issues if used on multiple machines.